Refactor compatibility modules into xarray.compat package (#10131)
Start of #10089

Move compatibility-related modules from xarray.core to new xarray.compat package:
- array_api_compat.py
- dask_array_compat.py
- dask_array_ops.py
- npcompat.py
- pdcompat.py
