+++ xarray/core/computation.py
@@ -22,9 +22,9 @@ from typing import TYPE_CHECKING, Any, Literal, TypeVar, Union, cast, overload
 
 import numpy as np
 
+from xarray.compat.array_api_compat import to_like_array
 from xarray.core import dtypes, duck_array_ops, utils
 from xarray.core.alignment import align, deep_align
-from xarray.core.array_api_compat import to_like_array
 from xarray.core.common import zeros_like
 from xarray.core.duck_array_ops import datetime_to_numeric
 from xarray.core.formatting import limit_lines
