+++ xarray/core/variable.py
@@ -16,9 +16,9 @@ from numpy.typing import ArrayLike
 from pandas.api.types import is_extension_array_dtype
 
 import xarray as xr  # only for Dataset and DataArray
+from xarray.compat.array_api_compat import to_like_array
 from xarray.core import common, dtypes, duck_array_ops, indexing, nputils, ops, utils
 from xarray.core.arithmetic import VariableArithmetic
-from xarray.core.array_api_compat import to_like_array
 from xarray.core.common import AbstractArray
 from xarray.core.extension_array import PandasExtensionArray
 from xarray.core.indexing import (
